mcu-sbMb-user

+++ build-scripts/4-whenWM/5-LibreOfficeConfig.sh
@@ -5,21 +5,21 @@
 
 set -ev
 
-# LibreOffice configurations
+# LibreOffice configurations ($CrPl/LibreOffice/Buildup.txt)
 
 #=> 0 config directory
 uc0=~/.config/libreoffice/4/user
 
 #=> 1 import
-rm -r $uc0/config
-rm $uc0/registrymodifications.xcu
-uc1=$CrPl/LibreOffice/lbg-sbMb-user
+[ -d "$uc0/config" ] && rm -r $uc0/config
+[ -s "$uc0/registrymodifications.xcu" ] && rm $uc0/registrymodifications.xcu
+uc1=$CrPl/LibreOffice/mcu-sbMb-user
 cp -r $uc1/config $uc0/config
 cp $uc1/registrymodifications.xcu $uc0/registrymodifications.xcu
 
-# #=> 1 export
-# uc1=$CrPl/LibreOffice/lbg-sbMb-user
-# mkdir $uc1
-# cp -r $uc0/config $uc1/config
-# cp $uc0/registrymodifications.xcu $uc1/registrymodifications.xcu
+#=> 1 export
+uc1=$CrPl/LibreOffice/lcu-sbMb-user
+mkdir $uc1
+cp -r $uc0/config $uc1/config
+cp $uc0/registrymodifications.xcu $uc1/registrymodifications.xcu
